Test principal funcionando
Solo se probo todas las condiciones con el criterioValorPrimeroEgreso, falta probar con los demas criterios. Agregue unos setters para las listas en vinculador. Elimine imports innecesarios. Arregle bug de criterio.termino(), habia que borrar el -1 de size

+++ Vinculacion I-E/Proceso vinculacion I-E/src/main/java/Domain/Entities/Condiciones/CondicionEntreFechas.java
@@ -7,8 +7,8 @@ import java.time.LocalDate;
 
 public class CondicionEntreFechas implements Condicion{
 
-    private int diasDesde;
-    private int diasHasta;
+    private long diasDesde;
+    private long diasHasta;
 
     @Override
     public boolean cumpleCondicion(Egreso egreso, Ingreso ingreso) {
